+++ solvers/bj_newton_pc.h
@@ -23,16 +23,16 @@ newton_pc_t* bj_newton_pc_new(void* context,
                               void (*compute_diag_block)(void* context, int i, real_t alpha, real_t beta, real_t gamma, real_t t, real_t* x, real_t* x_dot, real_t* diag_block),
                               void (*dtor)(void* context),
                               newton_pc_side_t side,
-                              int num_block_rows,
-                              int block_size);
+                              size_t num_block_rows,
+                              size_t block_size);
  
 // Creates a block Jacobi Newton preconditioner with variable block sizes.
 newton_pc_t* var_bj_newton_pc_new(void* context,
                                   void (*compute_diag_block)(void* context, int i, real_t alpha, real_t beta, real_t gamma, real_t t, real_t* x, real_t* x_dot, real_t* diag_block),
                                   void (*dtor)(void* context),
                                   newton_pc_side_t side,
-                                  int num_block_rows,
-                                  int* block_sizes);
+                                  size_t num_block_rows,
+                                  size_t* block_sizes);
 
 // Creates a block Jacobi Newton preconditioner that uses Curtis, Powell, and 
 // Reed's method (On the estimation of sparse Jacobian Matrices, 
@@ -45,9 +45,9 @@ newton_pc_t* cpr_bj_newton_pc_new(MPI_Comm comm,
                                   void (*dtor)(void* context),
                                   newton_pc_side_t side,
                                   adj_graph_t* sparsity,
-                                  int num_local_block_rows,
-                                  int num_remote_block_rows,
-                                  int block_size);
+                                  size_t num_local_block_rows,
+                                  size_t num_remote_block_rows,
+                                  size_t block_size);
 
 // This is the variable-block-size version of the Curtis-Powell-Reed block 
 // Jacobi preconditioner described above.
@@ -57,9 +57,9 @@ newton_pc_t* var_cpr_bj_newton_pc_new(MPI_Comm comm,
                                       void (*dtor)(void* context),
                                       newton_pc_side_t side,
                                       adj_graph_t* sparsity,
-                                      int num_local_block_rows,
-                                      int num_remote_block_rows,
-                                      int* block_sizes);
+                                      size_t num_local_block_rows,
+                                      size_t num_remote_block_rows,
+                                      size_t* block_sizes);
 
 // Creates a block Jacobi Newton preconditioner that uses Curtis, Powell, and 
 // Reed's method (On the estimation of sparse Jacobian Matrices, 
@@ -72,9 +72,9 @@ newton_pc_t* dae_cpr_bj_newton_pc_new(MPI_Comm comm,
                                       int (*F)(void* context, real_t t, real_t* x, real_t* xdot, real_t* Fval),
                                       void (*dtor)(void* context),
                                       adj_graph_t* sparsity,
-                                      int num_local_block_rows,
-                                      int num_remote_block_rows,
-                                      int block_size);
+                                      size_t num_local_block_rows,
+                                      size_t num_remote_block_rows,
+                                      size_t block_size);
 
 // This is the variable-block-size version of the Differential Algebraic 
 // Equation (DAE) Curtis-Powell-Reed block Jacobi preconditioner described above.
@@ -83,9 +83,9 @@ newton_pc_t* var_dae_cpr_bj_newton_pc_new(MPI_Comm comm,
                                           int (*F)(void* context, real_t t, real_t* x, real_t* xdot, real_t* Fval),
                                           void (*dtor)(void* context),
                                           adj_graph_t* sparsity,
-                                          int num_local_block_rows,
-                                          int num_remote_block_rows,
-                                          int* block_sizes);
+                                          size_t num_local_block_rows,
+                                          size_t num_remote_block_rows,
+                                          size_t* block_sizes);
 
 #endif
